% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphR.R
\name{GraphR_est}
\alias{GraphR_est}
\title{GraphR Model Estimation}
\usage{
GraphR_est(
  features,
  external,
  a_pi = 1,
  b_pi = 4,
  a_tau = 0.005,
  b_tau = 0.005,
  max_iter = 2000,
  max_tol = 0.001
)
}
\arguments{
\item{features}{Nodes of the graphs among which edges are built (e.g. a gene expression matrix of dimensions n \eqn{x} p).

Note: Please standardize each features before plugging into the estimation function.}

\item{external}{External covariates (n \eqn{x} q matrix).

Note: Please standardize continuous external covariates before plug into the
estimation function.}

\item{a_pi, b_pi}{\eqn{\pi} ~ Beta(\eqn{a_\pi, b_\pi})}

\item{a_tau, b_tau}{\eqn{\tau} ~ Gamma(\eqn{a_\tau, b_\tau})}

\item{max_iter}{maximum iterations}

\item{max_tol}{maximum tolerance}
}
\value{
\item{beta}{A p \eqn{x} p \eqn{x} q array of coefficients for external
covariates. The \eqn{[i,j,k]} element represents the effect of k-th
external covariates on regression of j-th node on i-th node.}
\item{phi}{A p \eqn{x} p \eqn{x} q array storing posterior inclusion probability (PIP)
of external covariates. The \eqn{[i,j,k]} elements represents the PIP of k-th
external covariates on regression of j-th node on i-th node.}
\item{omega_diag}{A p vector with i-th element representing the inverse
variance of error.}
}
\description{
Estimate the graphical regression coefficients and inclusion probabilities of external
covariates for the GraphR models.
}
\examples{
set.seed(100)
data("pam50")

features <- apply(pam50$features,2,scale) \%>\% as.matrix()
dim(features)
features[c(1:5),c(1:5)]

external <- pam50$external \%>\% as.matrix()
dim(external)
external[c(1:5),]


system.time(res <- GraphR_est(
  features,
  external,
  a_pi = 1,
  b_pi = 4,
  a_tau = 0.005,
  b_tau = 0.005,
  max_iter = 2000,
  max_tol = 0.001
))

}
