% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphR.R
\name{GraphR_visualization}
\alias{GraphR_visualization}
\title{Network Visualization.}
\usage{
GraphR_visualization(
  new_vec,
  graphR_est_res = NULL,
  beta = NULL,
  phi = NULL,
  omega_diag = NULL,
  fdr_thre = 0.01,
  magnitude_thre = 0.4
)
}
\arguments{
\item{new_vec}{A vector of new external covarites based on which plots
are made.

Note: Please ensure that the order and scale of new external covariates are same as those used in the estimation.}

\item{graphR_est_res}{Results from `GraphR_est` function.
If graphR_est_res = NULL, then beta, phi and omega_diag are needed simultaneously.}

\item{beta}{A p \eqn{x} p \eqn{x} q array storing coefficients of external
covariates. The \eqn{[i,j,k]} elements represents the effect of k-th
external covariates on regression of j-th node on i-th node.}

\item{omega_diag}{A p vector with i-th element representing the inverse
variance of error.}

\item{pip}{A p \eqn{x} p \eqn{x} q array storing posterior inclusion probability (PIP)
of external covariates. The \eqn{[i,j,k]} elements represents the PIP of k-th
external covariates on regression of j-th node on i-th node.}
}
\value{
\item{Plot}{Plot of circular networks. Node sizes represent connectivity degrees
of the corresponding features while edge widths are proportional to the partial
correlation between two features. Sign of the partial correlations are represented
by the color}
}
\description{
Plot of networks based on the given new external covariates vector
and thresholds for FDR-p values and magnitudes of partial correlation.
}
\examples{
data("Pam50")
library(magrittr)
features <- apply(Pam50$features,2,scale) \%>\% as.matrix()
features[c(1:5),c(1:5)]

external <- Pam50$external \%>\% as.matrix()
external[c(1:5),]


res <- GraphR_est(
  features,
  external,
  a_pi = 1,
  b_pi = 4,
  a_tau = 0.005,
  b_tau = 0.005,
  max_iter = 5,
  max_tol = 0.001
)

new_vec <- c(1,0,0)


GraphR_visualization(new_vec, graphR_est_res = res)





}
