% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pam50}
\alias{Pam50}
\title{PAM50 Molecular Intrinsic Subtypes Based Protemoics data set}
\format{
An object of class \code{list} of length 2.
}
\usage{
Pam50
}
\description{
This data set contains 190 protein with abundance information from 859 breast
cancer (BRCA) patients from The Cancer Genome Atlas (TCGA). Normal-like BRCA
are excluded due to the small sample sizes. Proteomics data are obtained and normalized from
The Cancer Proteome Atlas (TCPA) which applied Reverse Phase Protein
Arrays (RPPAs) to quantify proteins or phosphoproteins. PAM50 subtypes are treated
as external covariates. Notably, two subtypes, Luminal A and Luminal B BRCA, are
combined in one category to achieve higher power. That leads to three groups
with 626 BRCA patients in Luminal A and B, 75 patients in Her2-enriched and
158 in Basal-like.
}
\keyword{datasets}
