% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pan_gynae}
\alias{Pan_gynae}
\title{Proteomics data sets for Gynecological and Breast Cancers}
\format{
An object of class \code{list} of length 2.
}
\usage{
Pan_gynae
}
\description{
This dataset contains proteomics data for breast cancer (BRCA) and a group of
gynecological cancers including cervical squamous cell carcinoma
and endocervical adenocarcinoma (CESC), ovarian serous cystadenocarcinoma (OV)
and uterine corpus endometrial carcinoma (UCEC). Abundance of 189 kinds of
protein were measured across 1,941 patients among which 892 patients had BRCA,
173 had CESC, 436 had OV and 440 had UCEC.
}
\keyword{datasets}
