% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pct_repondants}
\alias{pct_repondants}
\title{pct_repondants}
\usage{
pct_repondants(repondants, total)
}
\arguments{
\item{table}{\dots}

\item{stats}{\dots}
}
\description{
pct_repondants
}
\keyword{internal}
