% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{quali_bi}
\alias{quali_bi}
\title{quali_bi}
\usage{
quali_bi(champ_quali, champ_valeur, identifiant, taille_texte = 3,
  taille_texte_legende = 1, orientation = "horizontal",
  label_pourcentage = FALSE, pct_suffix = "<U+202F>\%")
}
\arguments{
\item{champ_quali}{\dots}

\item{champ_valeur}{\dots}

\item{identifiant}{\dots}

\item{taille_texte}{\dots}

\item{taille_texte_legende}{\dots}

\item{orientation}{\dots}

\item{label_pourcentage}{\dots}

\item{pct_suffix}{\dots}
}
\description{
quali_bi
}
