% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{quali_bi_aires}
\alias{quali_bi_aires}
\title{quali_bi_aires}
\usage{
quali_bi_aires(champ_quali, champ_x, identifiant,
  label_pourcentage = FALSE, position_legende = "bas",
  taille_texte_legende = 1, nombre_lignes_legende = NULL,
  palette_ordinal = FALSE, pct_suffix = "<U+202F>\%")
}
\arguments{
\item{champ_quali}{\dots}

\item{champ_x}{\dots}

\item{identifiant}{\dots}

\item{label_pourcentage}{\dots}

\item{position_legende}{\dots}

\item{taille_texte_legende}{\dots}

\item{nombre_lignes_legende}{\dots}

\item{palette_ordinal}{\dots}

\item{pct_suffix}{\dots}
}
\description{
quali_bi_aires
}
