% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{quali_bi_ordinal}
\alias{quali_bi_ordinal}
\title{quali_bi_ordinal}
\usage{
quali_bi_ordinal(champ_quali, champ_valeur, identifiant, taille_texte = 3,
  taille_texte_legende = 1, orientation = "horizontal",
  label_pourcentage = FALSE)
}
\arguments{
\item{champ_quali}{\dots}

\item{champ_valeur}{\dots}

\item{identifiant}{\dots}

\item{taille_texte}{\dots}

\item{taille_texte_legende}{\dots}

\item{orientation}{\dots}

\item{label_pourcentage}{\dots}
}
\description{
quali_bi_ordinal
}
