% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{quali_uni}
\alias{quali_uni}
\title{quali_uni}
\usage{
quali_uni(
  champ_quali,
  lib_pct = TRUE,
  max_modalites = NULL,
  lib_modalite_autre = NULL,
  choix_multiple = FALSE,
  marge_gauche = FALSE,
  taille_texte = 3.5,
  ...
)
}
\arguments{
\item{champ_quali}{\dots}

\item{lib_pct}{\dots}

\item{max_modalites}{\dots}

\item{lib_modalite_autre}{\dots}

\item{choix_multiple}{\dots}

\item{marge_gauche}{\dots}

\item{taille_texte}{\dots}

\item{\dots}{\code{scales::percent} additional arguments}
}
\description{
quali_uni
}
