% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{quali_uni_aires}
\alias{quali_uni_aires}
\title{quali_uni_aires}
\usage{
quali_uni_aires(champ_x, identifiant, n_graph, n_population,
  label_pourcentage = FALSE, label_pourcentage_saut_ligne = TRUE,
  pct_suffix = "<U+202F>\%")
}
\arguments{
\item{identifiant}{\dots}

\item{n_graph}{\dots}

\item{n_population}{\dots}

\item{label_pourcentage}{\dots}

\item{label_pourcentage_saut_ligne}{\dots}

\item{pct_suffix}{\dots}

\item{table}{\dots}
}
\description{
quali_uni_aires
}
