% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{quali_uni_secteurs}
\alias{quali_uni_secteurs}
\title{quali_uni_secteurs}
\usage{
quali_uni_secteurs(champ_quali, max_modalites = NULL,
  marge_gauche = FALSE, effectif = TRUE, taille_texte = 3.5,
  marges = TRUE, ...)
}
\arguments{
\item{champ_quali}{\dots}

\item{max_modalites}{\dots}

\item{marge_gauche}{\dots}

\item{effectif}{\dots}

\item{taille_texte}{\dots}

\item{marges}{\dots}

\item{\dots}{\code{caractr::str_percent} additional arguments}
}
\description{
quali_uni_secteurs
}
