% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_areas_evolution}
\alias{shiny_areas_evolution}
\title{shiny_areas_evolution}
\usage{
shiny_areas_evolution(var_x, var_y, colors = NULL, title_x = "",
  title_y = "")
}
\arguments{
\item{var_x}{\dots}

\item{var_y}{\dots}

\item{colors}{\dots}

\item{title_x}{\dots}

\item{title_y}{\dots}
}
\description{
shiny_areas_evolution
}
