% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_barplot_vertical}
\alias{shiny_barplot_vertical}
\title{shiny_barplot_horizontal}
\usage{
shiny_barplot_vertical(var, colors = NULL, alpha = 1)
}
\arguments{
\item{var}{\dots}

\item{colors}{\dots}

\item{alpha}{\dots}
}
\description{
shiny_barplot_horizontal
}
