% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{shiny_colors}
\alias{shiny_colors}
\title{Palette de couleurs qualitative}
\usage{
shiny_colors(n)
}
\arguments{
\item{n}{Le nombre de couleurs désirées dans la palette.}
}
\value{
Une palette de couleur.
}
\description{
Renvoie une palette de couleur de 1 à 11 couleurs qualitative
}
