% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_donut}
\alias{shiny_donut}
\title{shiny_donut}
\usage{
shiny_donut(var, title = "", colors = NULL, alpha = 1)
}
\arguments{
\item{var}{\dots}

\item{title}{\dots}

\item{colors}{\dots}

\item{alpha}{\dots}
}
\description{
shiny_donut
}
