% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_line_base100}
\alias{shiny_line_base100}
\title{shiny_line_base100}
\usage{
shiny_line_base100(
  var_year,
  var_value,
  title_x = "",
  title_y = "",
  note_base100 = ""
)
}
\arguments{
\item{var_year}{\dots}

\item{var_value}{\dots}

\item{title_x}{\dots}

\item{title_y}{\dots}

\item{note_base100}{\dots}
}
\description{
shiny_line_base100
}
