% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_line_percent}
\alias{shiny_line_percent}
\title{shiny_line_percent}
\usage{
shiny_line_percent(var_year, var_percent, title_x = "", title_y = "",
  hovertext = NULL)
}
\arguments{
\item{var_year}{\dots}

\item{var_percent}{\dots}

\item{title_x}{\dots}

\item{title_y}{\dots}

\item{hovertext}{\dots}
}
\description{
shiny_line_percent
}
