% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_line_percent_multi}
\alias{shiny_line_percent_multi}
\title{shiny_line_percent}
\usage{
shiny_line_percent_multi(var_year, var_line, var_percent, title_x = "",
  title_y = "")
}
\arguments{
\item{var_year}{\dots}

\item{var_percent}{\dots}

\item{title_x}{\dots}

\item{title_y}{\dots}
}
\description{
shiny_line_percent
}
