% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_pie}
\alias{shiny_pie}
\title{shiny_pie}
\usage{
shiny_pie(
  var,
  colors = NULL,
  alpha = 1,
  donut = FALSE,
  donut_title = "",
  legend_position = c("right", "bottom")
)
}
\arguments{
\item{var}{\dots}

\item{colors}{\dots}

\item{alpha}{\dots}

\item{donut}{\dots}

\item{donut_title}{\dots}

\item{legend_position}{\dots}
}
\description{
shiny_pie
}
