% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_treemap}
\alias{shiny_treemap}
\title{shiny_treemap}
\usage{
shiny_treemap(var_x, colors = NULL, alpha = 1)
}
\arguments{
\item{var_x}{\dots}

\item{colors}{\dots}

\item{alpha}{\dots}
}
\description{
shiny_treemap
}
