% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Si}
\alias{Si}
\title{Proteomics data with Stemness Indices and Age in Breast Cancer}
\format{
An object of class \code{list} of length 2.
}
\usage{
Si
}
\description{
189 protein abundance measurement across 616 breast cancer (BRCA) patients
from The Cancer Genome Atlas (TCGA) are included and treated as features.
External covariates include two independent stemness indices (SI) which are
obtained based on DNA methylation (mDNAsi) and (mRNAsi), and age of patients
The mDNAsi and mRNAsi provide the degrees of dedifferentiation on epigenetic and
gene expression level respectively, which range from 0 to 1 with lower values
implying tendency to normal-like cells.
}
\keyword{datasets}
