% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial}
\alias{spatial}
\title{Spatial transcriptomics data set for breast cancer}
\format{
An object of class \code{list} of length 3.
}
\usage{
spatial
}
\description{
This data set was collected from biopsy of breast cancer at
a thickness of 16 $\mu$m whose locations can be classified into
three spatial regions as tumor, intermediate, and normal based on the
Hematoxylin and Eosin (H\&E) staining image. The size of three regions are
114, 67, and 69 respectively. The data includes measurement of 100 spatially expressed
genes with the lowest Benjamini-Hochberg (BH) adjusted p-value by applying SPARK method
at 250 spot locations.
}
\keyword{datasets}
