% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stats_count_bi}
\alias{stats_count_bi}
\title{stats_count_bi}
\usage{
stats_count_bi(champ_quali, champ_x, identifiant = NULL, complet = FALSE)
}
\arguments{
\item{champ_quali}{\dots}

\item{champ_x}{\dots}

\item{identifiant}{\dots}

\item{complet}{\dots}
}
\description{
stats_count_bi
}
