% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stats_count_uni}
\alias{stats_count_uni}
\title{stats_count_uni}
\usage{
stats_count_uni(champ_quali, max_modalites = NULL,
  lib_modalite_autre = NULL, choix_multiple = FALSE)
}
\arguments{
\item{champ_quali}{\dots}

\item{max_modalites}{\dots}

\item{lib_modalite_autre}{\dots}

\item{choix_multiple}{\dots}
}
\description{
stats_count_uni
}
\keyword{internal}
