% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{update_lambda_r}
\alias{update_lambda_r}
\title{wrap function}
\usage{
update_lambda_r(n, e_zbeta_sq, len_res_sq)
}
\arguments{
\item{n}{sample size}

\item{e_zbeta_sq}{expectation}

\item{len_res_sq}{length}
}
\description{
wrap function
}
