% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphR.R
\name{GraphR_est}
\alias{GraphR_est}
\title{GraphR Model Estimation}
\usage{
GraphR_est(
  features,
  external,
  a_pi = 1,
  b_pi = 4,
  a_tau = 0.005,
  b_tau = 0.005,
  max_iter = 2000,
  max_tol = 0.001
)
}
\arguments{
\item{features}{Known as nodes in graphs among which connections are built.
A \eqn{n \times p} matrix.}

\item{external}{External covariates. A \eqn{n \times q} matrix.}

\item{a_pi, b_pi}{\eqn{\pi \sim} Beta (\eqn{a_\pi,b_\pi})}

\item{a_tau, b_tau}{\eqn{\tau \sim} Gamma (\eqn{a_\tau,b_\tau})}

\item{max_iter}{maximum iterations}

\item{max_tol}{maximum tolerance}
}
\value{
1. beta: A \eqn{p \times p \times q} array storing coefficients of external
covariates. The \eqn{[i,j,k]} elements represents the effect of \eqn{k^{th}}
external covariates on regression of \eqn{j^{th}} node on \eqn{i^{th}} node

2. phi: A \eqn{p \times p \times q} array storing posterior inclusion probability (PIP)
of external covariates. The \eqn{[i,j,k]} elements represents the PIP of \eqn{k^{th}}
external covariates on regression of \eqn{j^{th}} node on \eqn{i^{th}} node

3. omega_diag: A p vector with \eqn{i^{th}} element representing the inverse
variance of error
}
\description{
Estimate coefficients and inclusion probabilities of external
covariates in GraphR models
}
