% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphR.R
\name{GraphR_pred}
\alias{GraphR_pred}
\title{GraphR Model Predictions}
\usage{
GraphR_pred(
  new_df,
  graphR_est_res = NULL,
  beta = NULL,
  phi = NULL,
  omega_diag = NULL
)
}
\arguments{
\item{new_df}{A matrix of new external covarites based on which predictions
are made.

Note: Please ensure that the order and scale of new external covariates are same as those used in the estimation.}

\item{graphR_est_res}{Results from `GraphR_est` function.
If graphR_est_res = NULL, then beta, phi and omega_diag are needed simultaneously.}

\item{beta}{A p \eqn{x} p \eqn{x} q array storing coefficients of external
covariates. The \eqn{[i,j,k]} elements represents the effect of k-th
external covariates on regression of j-th node on i-th node.}

\item{omega_diag}{A p vector with i-th element representing the inverse
variance of error.}

\item{pip}{A p \eqn{x} p \eqn{x} q array storing posterior inclusion probability (PIP)
of external covariates. The \eqn{[i,j,k]} elements represents the PIP of k-th
external covariates on regression of j-th node on i-th node.}
}
\value{
\item{feature_id1, feature_id2}{Indices of nodes.}

\item{Pr_inclusion}{Posterior inclusion probability of connections between two nodes
based on "And" rules.}

\item{Correlation}{Partial correlation between two nodes. Values with maximum magnitudes are provided.}

\item{FDR_p}{Bayesian FDR-adjusted p values.}
}
\description{
Prediction of partial correlation between two nodes and the
corresponding inclusion probabilities from the results of GraphR model along with
Bayesian FDR-adjusted p-values.
}
\examples{
set.seed(100)
data("Pam50")
library(magrittr)
features <- apply(Pam50$features,2,scale) \%>\% as.matrix()
features[c(1:5),c(1:5)]

external <- Pam50$external \%>\% as.matrix()
external[c(1:5),]


system.time(res <- GraphR_est(
  features,
  external,
  a_pi = 1,
  b_pi = 4,
  a_tau = 0.005,
  b_tau = 0.005,
  max_iter = 5,
  max_tol = 0.001
))

####### prediction
new_df <- diag(3)
colnames(new_df) <- colnames(external)

pred <- GraphR_pred(new_df, res)
head(pred)





}
