% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphR.R
\name{graphR_pred}
\alias{graphR_pred}
\title{GraphR Model Predictions}
\usage{
graphR_pred(
  new_df,
  graphR_est_res = NULL,
  beta = NULL,
  phi = NULL,
  omega_diag = NULL
)
}
\arguments{
\item{new_df}{A matrix of new external covarites based on which predicitons
are made}

\item{graphR_est_res}{Results of `graphR_est`}

\item{beta:}{A \eqn{p \times p \times q} array storing coefficients of external
covariates. The \eqn{[i,j,k]} elements represents the effect of \eqn{k^{th}}
external covariates on regression of \eqn{j^{th}} node on \eqn{i^{th}} node}

\item{A}{\eqn{p \times p \times q} array storing posterior inclusion probability (PIP)
of external covariates. The \eqn{[i,j,k]} elements represents the PIP of \eqn{k^{th}}
external covariates on regression of \eqn{j^{th}} node on \eqn{i^{th}} node}

\item{omega_diag:}{A p vector with \eqn{i^{th}} element representing the inverse
variance of error}
}
\value{
1. feature_id1, feature_id2: Iindices of nodes

2. Pr_inclusion: Posterior inclusion probability of connections between two nodes
based on "And" rules.

3. Correlation: Partial correlation between two nodes. Values with maximum magnitudes are provided

4. FDR_p: Bayesian FDR-adjusted p values
}
\description{
Predict partial correlation between two nodes and the
corresponding inclusion probabilities from the results of GraphR model.
Bayesian FDR-adjusted p-values are given.
}
