% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphR.R
\name{GraphR_pred}
\alias{GraphR_pred}
\title{GraphR Model Predictions}
\usage{
GraphR_pred(
  new_df,
  graphR_est_res = NULL,
  beta = NULL,
  phi = NULL,
  omega_diag = NULL
)
}
\arguments{
\item{new_df}{A matrix of new external covarites based on which predicitons
are made

Note: Please standardize continuous external covariates before plug into the
estimation function}

\item{graphR_est_res}{Results of `graphR_est`}

\item{beta:}{A \eqn{p \times p \times q} array storing coefficients of external
covariates. The \eqn{[i,j,k]} elements represents the effect of \eqn{k^{th}}
external covariates on regression of \eqn{j^{th}} node on \eqn{i^{th}} node}

\item{pip}{\eqn{p \times p \times q} array storing posterior inclusion probability (PIP)
of external covariates. The \eqn{[i,j,k]} elements represents the PIP of \eqn{k^{th}}
external covariates on regression of \eqn{j^{th}} node on \eqn{i^{th}} node}

\item{omega_diag:}{A p vector with \eqn{i^{th}} element representing the inverse
variance of error}
}
\value{
1. feature_id1, feature_id2: Indices of nodes

2. Pr_inclusion: Posterior inclusion probability of connections between two nodes
based on "And" rules.

3. Correlation: Partial correlation between two nodes. Values with maximum magnitudes are provided

4. FDR_p: Bayesian FDR-adjusted p values
}
\description{
Predict partial correlation between two nodes and the
corresponding inclusion probabilities from the results of GraphR model.
Bayesian FDR-adjusted p-values are given.
}
\examples{
set.seed(100)
data("pam50")

features <- apply(pam50$features,2,scale) \%>\% as.matrix()
features[c(1:5),c(1:5)]

external <- pam50$external \%>\% as.matrix()
external[c(1:5),]


system.time(res <- GraphR_est(
  features,
  external,
  a_pi = 1,
  b_pi = 4,
  a_tau = 0.005,
  b_tau = 0.005,
  max_iter = 2000,
  max_tol = 0.001
))

####### prediction
new_df <- diag(3)
colnames(new_df) <- colnames(external)

pred <- GraphR_pred(new_df, res)
head(pred)





}
