% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{quali_bi}
\alias{quali_bi}
\title{quali_bi}
\usage{
quali_bi(
  champ_quali,
  champ_valeur,
  identifiant,
  taille_texte = 3,
  taille_texte_legende = 1,
  orientation = "horizontal",
  label_pourcentage = FALSE,
  ...
)
}
\arguments{
\item{champ_quali}{\dots}

\item{champ_valeur}{\dots}

\item{identifiant}{\dots}

\item{taille_texte}{\dots}

\item{taille_texte_legende}{\dots}

\item{orientation}{\dots}

\item{label_pourcentage}{\dots}

\item{\dots}{\code{scales::percent} additional arguments}
}
\description{
quali_bi
}
