% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{quali_bi_aires2}
\alias{quali_bi_aires2}
\title{quali_bi_aires2}
\usage{
quali_bi_aires2(
  champ_quali,
  champ_x,
  label_effectif = FALSE,
  position_legende = "bas",
  taille_texte_legende = 1,
  taille_texte_axe_x = 9,
  nombre_lignes_legende = NULL,
  palette_ordinal = FALSE,
  ...
)
}
\arguments{
\item{champ_quali}{\dots}

\item{champ_x}{\dots}

\item{label_effectif}{\dots}

\item{position_legende}{\dots}

\item{taille_texte_legende}{\dots}

\item{taille_texte_axe_x}{\dots}

\item{nombre_lignes_legende}{\dots}

\item{palette_ordinal}{\dots}

\item{\dots}{\code{scales::percent} additional arguments}
}
\description{
quali_bi_aires2
}
