% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_areas_evolution}
\alias{shiny_areas_evolution}
\title{shiny_areas_evolution}
\usage{
shiny_areas_evolution(
  var_x,
  var_y,
  colors = NULL,
  title_x = "",
  title_y = "",
  font_family = NULL
)
}
\arguments{
\item{var_x}{\dots}

\item{var_y}{\dots}

\item{colors}{\dots}

\item{title_x}{\dots}

\item{title_y}{\dots}

\item{font_family}{\dots}
}
\description{
shiny_areas_evolution
}
