% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_barplot_horizontal_multi}
\alias{shiny_barplot_horizontal_multi}
\title{shiny_barplot_horizontal_multi}
\usage{
shiny_barplot_horizontal_multi(var_x, var_y, colors = NULL, alpha = 1,
  title_x = "", title_y = "")
}
\arguments{
\item{var_x}{\dots}

\item{var_y}{\dots}

\item{colors}{\dots}

\item{alpha}{\dots}

\item{title_x}{\dots}

\item{title_y}{\dots}
}
\description{
shiny_barplot_horizontal_multi
}
