% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_treemap_bi}
\alias{shiny_treemap_bi}
\title{shiny_treemap_bi}
\usage{
shiny_treemap_bi(parents, labels, colors = NULL, alpha = 1)
}
\arguments{
\item{parents}{\dots}

\item{labels}{\dots}

\item{colors}{\dots}

\item{alpha}{\dots}
}
\description{
shiny_treemap_bi
}
