% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stats_count_uni}
\alias{stats_count_uni}
\title{stats_count_uni}
\usage{
stats_count_uni(champ_quali, choix_multiple_labels = NULL)
}
\arguments{
\item{champ_quali}{\dots}

\item{choix_multiple_labels}{\dots}
}
\description{
stats_count_uni
}
